move `useWindowDimensions` to separated file
