+++ Composer/packages/extensions/visual-designer/src/utils/hooks.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { useContext, useState, useEffect, useRef } from 'react';
-import debounce from 'lodash/debounce';
+import { useContext, useState, useEffect } from 'react';
 import { LgTemplateRef } from '@bfc/shared';
 
 import { NodeRendererContext } from '../store/NodeRendererContext';
@@ -53,24 +52,4 @@ export const useLgTemplate = (str?: string, dialogId?: string) => {
   });
 
   return templateText;
-};
-
-const getWindowDimensions = () => {
-  const { innerWidth: width, innerHeight: height } = window;
-  return {
-    width,
-    height,
-  };
-};
-
-export const useWindowDimensions = () => {
-  const [windowDimensions, setWindowDimensions] = useState(getWindowDimensions());
-  const handleResize = useRef(debounce(() => setWindowDimensions(getWindowDimensions()), 200)).current;
-
-  useEffect(() => {
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  return windowDimensions;
 };
